chore: renamed the server helpers directory as utils and updated its imports accordingly
