+++ server/api/transactions.js
@@ -1,7 +1,7 @@
 // Imports
 const router = require('express').Router()
 
-const {quote, cents} = require('../helpers')
+const {quote, cents} = require('../utils')
 
 // Models
 const {Transaction, Portfolio} = require('../db/models')
